llm-chat/requirements.txt
# requirements.txt
streamlit>=1.31.0
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
redis>=5.0.1
python-dateutil>=2.8.2